Dummy Func and Binary Output
Implement dummy kernel function to calculate range, check against threshold value, and write to a bool array a 1 if range of macroblock is >= threshold.

Removed function implementations from header files. Previously there were errors when including headers in multiple files.

This will be what Martinez and Mykyta will work on for optimization before the dataset and models are complete.