DTC Serial Prototype Function Creation
Title.

OpenCL implementation in the current prototype is solely for statistic collection. DTC is implemented serially, using OpenCV.

Hyperparameter tuning is TODO, as large amount of false negatives are occurring, according to visual output.

Output was tested with an input value shown to result in a positive response from model, so pipeline is implemented correctly, but DTC is not.